Moved functions in the main pipeline to several modules
