+++ src/config/config.py
@@ -1,0 +1,3 @@
+WORDS_PLACE_IN_WORDS_COUNT_LIST = 0
+FREQUENCY_PLACE_IN_FREQUENCY_LIST = 1
+DEFAULT_COUNT_WHEN_WORD_IS_NON_EXISTING = 1
