+++ src/find_meaningful_words.py
@@ -1,111 +1,44 @@
-from collections import Counter
 from src.utils import SaveRead
-from src.count_words import get_words_only_small_letters, count_words_from_file
+from src import get_words_only, words_counter, ratio_of_words_count
 
-#
-WORDS_PLACE_IN_FREQUENCY_LIST = 0
-FREQUENCY_PLACE_IN_FREQUENCY_LIST = 1
 
 # TODO: Add documentation to each function.
 # TODO: Add tests to each function
 # TODO: Divide function into different files.
 # TODO: Add configuration file for magic numbers and strings.
 
+def count_words_from_file(file_path):
+    file_string = SaveRead.read_txt_file_to_string(file_path)
+    lower_letters_words_list = get_words_only.get_words_only_small_letters(file_string)
+    words_count = len(lower_letters_words_list) 
+    return words_count
 
-def sort_words_count(counted_words_tuple_list):
-    sorted_counted_words_tuple_list = sorted(
-        counted_words_tuple_list, key=lambda x: x[FREQUENCY_PLACE_IN_FREQUENCY_LIST], reverse=True)
-    return sorted_counted_words_tuple_list
-
-
-def count_words_in_list(words_list):
-    counted_words_dict = dict(Counter(words_list))
-    counted_words_tuple_list = counted_words_dict.items()
-    sorted_counted_words_tuple_list = sort_words_count(
-        counted_words_tuple_list)
-    return sorted_counted_words_tuple_list
-
-
-def get_only_common_words(words_frequencies_dict, number_of_most_common):
-    return [x[WORDS_PLACE_IN_FREQUENCY_LIST] for x in words_frequencies_dict[:number_of_most_common]]
 
 
 def get_most_common_words_in_string(file_path, number_of_most_common):
 
     file_string = SaveRead.read_txt_file_to_string(file_path)
-    only_words_list = get_words_only_small_letters(file_string)
-    counted_words_tuple_list = count_words_in_list(only_words_list)
-    most_common_words = get_only_common_words(
+    only_words_list = get_words_only.get_words_only_small_letters(file_string)
+    counted_words_tuple_list = words_counter.count_words_in_list(only_words_list)
+    most_common_words = words_counter.get_only_common_words(
         counted_words_tuple_list, number_of_most_common)
     return most_common_words
 
 
-def turn_words_count_tuple_list_to_words_only(frequencies_list):
-    return [x[WORDS_PLACE_IN_FREQUENCY_LIST] for x in frequencies_list]
-
-
-def add_missing_base_words_to_reference_words_count(base_words, reference_words, reference_frequencies):
-    base_words_not_in_ref = set(base_words) ^ set(reference_words)
-    reference_counted_words_and_base_words = reference_frequencies + \
-        [(x, 1) for x in base_words_not_in_ref]
-    return reference_counted_words_and_base_words
-
-
-def get_base_words_with_reference_count(base_words_list, reference_words_list, reference_words_count_tuple_list):
-    reference_words_count_with_base_words_addition = add_missing_base_words_to_reference_words_count(
-        base_words_list, reference_words_list, reference_words_count_tuple_list)
-    joint_reference_base_words_list = turn_words_count_tuple_list_to_words_only(
-        reference_words_count_with_base_words_addition)
-    # Find the index of words from base list to reference
-    base_words_index_in_joint_reference_words_list = [joint_reference_base_words_list.index(
-        x) for x in base_words_list]
-    # Get base words with reference count
-    base_words_with_reference_count = [reference_words_count_with_base_words_addition[x]
-                                       for x in base_words_index_in_joint_reference_words_list]
-    return base_words_with_reference_count
-
-
-def calculate_relative_words_count_of_2_lists(base_words_count_tuple_list, base_words_with_reference_count):
-    range_base_words = range(len(base_words_count_tuple_list))
-    # For each word in base list, calculate the relative count to the reference words list
-    base_relative_to_reference_count = [base_words_count_tuple_list[x][FREQUENCY_PLACE_IN_FREQUENCY_LIST] /
-                                        base_words_with_reference_count[x][FREQUENCY_PLACE_IN_FREQUENCY_LIST]
-                                        for x in range_base_words]
-    # Create a words_ count tuple list
-    base_words_relative_reference_count_tuple_list = [
-        (base_words_count_tuple_list[x][WORDS_PLACE_IN_FREQUENCY_LIST], 
-        base_relative_to_reference_count[x]) 
-        for x in range_base_words]
-    return base_words_relative_reference_count_tuple_list
-
-
-def get_ratio_of_base_words_count_relative_to_reference_words_count(base_words_count_tuple_list,
-                                                                    reference_words_count_tuple_list):
-    base_words_list = turn_words_count_tuple_list_to_words_only(
-        base_words_count_tuple_list)
-    reference_words_list = turn_words_count_tuple_list_to_words_only(
-        reference_words_count_tuple_list)
-    base_words_with_reference_count = get_base_words_with_reference_count(
-        base_words_list, reference_words_list, reference_words_count_tuple_list)
-    base_words_relative_reference_count_tuple_list = calculate_relative_words_count_of_2_lists(
-        base_words_count_tuple_list, base_words_with_reference_count)
-
-    return base_words_relative_reference_count_tuple_list
-
-
 def find_meaningful_words_in_file_from_another_file(base_path, reference_path, number_of_words):
     # Count words in base and reference files
-    base_words_counted = count_words_in_list(
-        get_words_only_small_letters(SaveRead.read_txt_file_to_string(base_path)))
-    reference_words_counted = count_words_in_list(
-        get_words_only_small_letters(SaveRead.read_txt_file_to_string(reference_path)))
+    base_words_counted = words_counter.count_words_in_list(
+        get_words_only.get_words_only_small_letters(SaveRead.read_txt_file_to_string(base_path)))
+    reference_words_counted = words_counter.count_words_in_list(
+        get_words_only.get_words_only_small_letters(SaveRead.read_txt_file_to_string(reference_path)))
     # Get the ratio of base words count relative to reference words count.
-    base_words_relative_reference_count_tuple_list = get_ratio_of_base_words_count_relative_to_reference_words_count(
+    base_words_relative_reference_count_tuple_list = \
+        ratio_of_words_count.get_ratio_of_base_words_count_relative_to_reference_words_count(
         base_words_counted, reference_words_counted)
     # Sort the relative count of base words
-    sorted_base_words_relative_reference_count_tuple_list = sort_words_count(
+    sorted_base_words_relative_reference_count_tuple_list = words_counter.sort_words_count(
         base_words_relative_reference_count_tuple_list)
-    most_meaningful_words = get_only_common_words(
+    most_meaningful_words = words_counter.get_only_common_words(
         sorted_base_words_relative_reference_count_tuple_list, number_of_words)
     return most_meaningful_words
